+++ otter/assign/assignment.py
@@ -104,6 +104,20 @@ class Assignment(fica.Config, Loggable):
             default=[],
         )
 
+        class RequireNoPDFAckValue(fica.Config):
+
+            message: str = fica.Key(
+                description="a message to show to students if a PDF is meant to be included in " \
+                    "the submission but cannot be generated",
+            )
+
+        require_no_pdf_ack: Union[bool, RequireNoPDFAckValue] = fica.Key(
+            description="whether to require students to acknowledge that a PDF could not be " \
+                "created if one is meant to be included in the submission zip file",
+            default=False,
+            subkey_container=RequireNoPDFAckValue,
+        )
+
     export_cell: ExportCellValue = fica.Key(
         description="whether to include an Otter export cell in the output notebooks",
         subkey_container=ExportCellValue,
@@ -244,9 +258,11 @@ class Assignment(fica.Config, Loggable):
         self._logger.debug(f"Initializing with config: {user_config}")
         super().__init__(user_config, **kwargs)
 
-        # convert a boolean to a config object for self.generate if indicated
+        # convert true values masking subkey contains to those containers
         if self.generate is True:
             self.generate = AutograderConfig()
+        if self.export_cell is True:
+            self.export_cell = type(self).ExportCellValue()
 
         if self.variables:
             warnings.warn(
@@ -258,6 +274,8 @@ class Assignment(fica.Config, Loggable):
         ret = super().update(user_config)
         if self.generate is True:
             self.generate = AutograderConfig()
+        if self.export_cell is True:
+            self.export_cell = type(self).ExportCellValue()
         return ret
 
     @property