+++ otter/assign/notebook_transformer.py
@@ -11,11 +11,24 @@ from .plugins import replace_plugins_with_calls
 from .question_config import QuestionConfig
 from .r_adapter import rmarkdown_converter
 from .r_adapter.cell_factory import RCellFactory
-from .solutions import has_seed, SOLUTION_CELL_TAG, overwrite_seed_vars, strip_ignored_lines, \
-    strip_solutions_and_output
+from .solutions import (
+    has_seed,
+    SOLUTION_CELL_TAG,
+    overwrite_seed_vars,
+    strip_ignored_lines,
+    strip_solutions_and_output,
+)
 from .tests_manager import AssignmentTestsManager
-from .utils import add_tag, add_assignment_name_to_notebook, AssignNotebookFormatException, \
-    get_source, is_cell_type, is_ignore_cell, lock, remove_cell_ids
+from .utils import (
+    add_tag,
+    add_assignment_name_to_notebook,
+    add_require_no_pdf_ack_to_notebook,
+    AssignNotebookFormatException,
+    get_source,
+    is_cell_type,
+    is_ignore_cell,
+    lock,
+)
 
 
 class NotebookTransformer:
@@ -109,10 +122,6 @@ class NotebookTransformer:
             transformed_cells += self.cell_factory.create_check_all_cells()
 
         if self.assignment.export_cell:
-            export_cell = self.assignment.export_cell
-            if export_cell is True:
-                export_cell = {}
-
             transformed_cells += self.cell_factory.create_export_cells()
 
         transformed_nb = copy.deepcopy(nb)
@@ -124,11 +133,9 @@ class NotebookTransformer:
         # strip out ignored lines
         transformed_nb = strip_ignored_lines(transformed_nb)
 
-        # TODO: this is a bad practice and only a monkey-patch for #340. we should do some better
-        # parsing of the nbformat version info to determine if this is necessary.
-        remove_cell_ids(transformed_nb)
-
         add_assignment_name_to_notebook(transformed_nb, self.assignment)
+
+        add_require_no_pdf_ack_to_notebook(transformed_nb, self.assignment)
 
         return TransformedNotebookContainer(transformed_nb, self)