+++ test/test_generate/files/autograder-r-requirements-correct/setup.sh
@@ -33,4 +33,4 @@ mamba run -n otter-env Rscript /autograder/source/requirements.r
 mamba init --all
 
 # install ottr
-mamba run -n otter-env Rscript -e 'install.packages("https://cran.r-project.org/package=ottr&version=1.2.0", dependencies=TRUE, repos=NULL)'
+mamba run -n otter-env Rscript -e 'install.packages("https://cran.r-project.org/package=ottr&version=1.4.0", dependencies=TRUE, repos=NULL)'